w3c-reports/12092014/mac/runned-tests/notifications/interfaces-expected.txt

CONSOLE MESSAGE: line 188: callback not yet supported
CONSOLE MESSAGE: line 188: callback not yet supported
CONSOLE MESSAGE: Deprecated attempt to access property 'onclick' on a non-Notification object.
CONSOLE MESSAGE: Deprecated attempt to access property 'onshow' on a non-Notification object.
CONSOLE MESSAGE: Deprecated attempt to access property 'onerror' on a non-Notification object.
CONSOLE MESSAGE: Deprecated attempt to access property 'onclose' on a non-Notification object.
CONSOLE MESSAGE: Deprecated attempt to access property 'dir' on a non-Notification object.
CONSOLE MESSAGE: Deprecated attempt to access property 'tag' on a non-Notification object.

FAIL Notification interface: existence and properties of interface object assert_equals: prototype of window's property "Notification" is not Function.prototype expected (function) function "function () {
    [native code]
}" but got (object) object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/resources/idlharness.js:991:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS Notification interface object length 
FAIL Notification interface: existence and properties of interface prototype object assert_false: Notification.prototype is enumerable expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/resources/idlharness.js:1094:21
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: existence and properties of interface prototype object's "constructor" property assert_true: Notification.prototype.constructor is not writable expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1183:20
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS Notification interface: attribute permission 
PASS Notification interface: operation requestPermission(NotificationPermissionCallback) 
FAIL Notification interface: attribute onclick assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onshow assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onerror assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute onclose assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute title assert_true: The prototype object must have a property "title" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute dir assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute lang assert_true: The prototype object must have a property "lang" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute body assert_true: The prototype object must have a property "body" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute tag assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: attribute icon assert_true: The prototype object must have a property "icon" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS Notification interface: operation close() 
PASS Notification must be primary interface of new Notification("foo") 
PASS Stringification of new Notification("foo") 
PASS Notification interface: new Notification("foo") must inherit property "permission" with the proper type (0) 
PASS Notification interface: new Notification("foo") must inherit property "requestPermission" with the proper type (1) 
PASS Notification interface: calling requestPermission(NotificationPermissionCallback) on new Notification("foo") with too few arguments must throw TypeError 
PASS Notification interface: new Notification("foo") must inherit property "onclick" with the proper type (2) 
PASS Notification interface: new Notification("foo") must inherit property "onshow" with the proper type (3) 
PASS Notification interface: new Notification("foo") must inherit property "onerror" with the proper type (4) 
PASS Notification interface: new Notification("foo") must inherit property "onclose" with the proper type (5) 
FAIL Notification interface: new Notification("foo") must inherit property "title" with the proper type (6) assert_inherits: property "title" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS Notification interface: new Notification("foo") must inherit property "dir" with the proper type (7) 
FAIL Notification interface: new Notification("foo") must inherit property "lang" with the proper type (8) assert_inherits: property "lang" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
FAIL Notification interface: new Notification("foo") must inherit property "body" with the proper type (9) assert_inherits: property "body" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS Notification interface: new Notification("foo") must inherit property "tag" with the proper type (10) 
FAIL Notification interface: new Notification("foo") must inherit property "icon" with the proper type (11) assert_inherits: property "icon" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS Notification interface: new Notification("foo") must inherit property "close" with the proper type (12) 
PASS EventTarget interface: new Notification("foo") must inherit property "addEventListener" with the proper type (0) 
FAIL EventTarget interface: calling addEventListener(DOMString,EventListener,boolean) on new Notification("foo") with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS EventTarget interface: new Notification("foo") must inherit property "removeEventListener" with the proper type (1) 
FAIL EventTarget interface: calling removeEventListener(DOMString,EventListener,boolean) on new Notification("foo") with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/notifications/interfaces.html:79:14)
PASS EventTarget interface: new Notification("foo") must inherit property "dispatchEvent" with the proper type (2) 
PASS EventTarget interface: calling dispatchEvent(Event) on new Notification("foo") with too few arguments must throw TypeError 
